Added files via upload
This is an analysis of 2016 Presidential Primary contributions made in Iowa. I explored the dataset, added new features, grouped the data in various ways, made plots, and finally synthesized my findings. As a teaser - Hillary Clinton and Ben Carson received the most contributions. Can you guess which demographic in Iowa contributes the most, or what city they live in? Check out the analysis to find out!